erased unnecessary components and finished metadata & og doc design
