+++ website/src/pages/documents/print-document.css.ts
@@ -1,18 +1,6 @@
 import { globalStyle, style } from "@vanilla-extract/css"
-import {
-  annotationSection,
-  documentBlock,
-  inlineBlock,
-  wordGroup,
-  wordGroupInline,
-} from "src/segment.css"
-import {
-  colors,
-  fonts,
-  hspace,
-  mediaQueries,
-  vspace,
-} from "src/style/constants"
+import { wordGroup, wordGroupInline } from "src/segment.css"
+import { fonts, hspace, mediaQueries, vspace } from "src/style/constants"
 
 const printFonts = {
   quattrocento: `"Quattrocento", serif`,
@@ -20,7 +8,6 @@ const printFonts = {
 
 const printFontSizes = {
   large: "1.1rem",
-  legendItem: "1rem",
 }
 
 export const printHeading = style({
@@ -97,77 +84,6 @@ export const printBlankPage = style({
       fontStyle: "italic",
       fontFamily: "serif",
       fontSize: "1rem",
-    },
-  },
-})
-
-export const printMetaSection = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "flex",
-      flexDirection: "column",
-      width: "100%",
-      marginTop: vspace.one,
-    },
-  },
-})
-
-export const printMetaField = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "flex",
-      flexDirection: "row",
-      paddingTop: vspace.quarter,
-      paddingBottom: vspace.quarter,
-      borderBottom: `1px solid #ADADAD`,
-    },
-  },
-})
-
-export const printMetaLabel = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "block",
-      width: "160px",
-      flexShrink: 0,
-      fontFamily: fonts.header,
-      fontSize: "10pt",
-      fontWeight: 600,
-      color: "#333333",
-    },
-  },
-})
-
-export const printMetaValue = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "block",
-      fontFamily: "serif",
-      fontSize: "10pt",
-      fontWeight: 400,
-      color: "#333333",
-      lineHeight: 1.4,
-    },
-  },
-})
-
-export const printImageSource = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "block",
-      fontFamily: "serif",
-      fontStyle: "normal",
-      fontWeight: 400,
-      fontSize: "14px",
-      lineHeight: 1.2,
-      width: "636px",
-      height: "17px",
-      textAlign: "center",
     },
   },
 })
@@ -260,142 +176,6 @@ globalStyle(`${printLegendBox} p`, {
     },
   },
 })
-
-export const printSectionHeading = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "flex",
-      flexDirection: "column",
-      alignItems: "center",
-      fontFamily: `"Quattrocento", serif`,
-      fontSize: "1.158rem",
-      fontWeight: 700,
-      lineHeight: 1.2,
-      marginTop: vspace.one,
-      marginBottom: 0,
-      breakBefore: "avoid",
-      breakAfter: "avoid",
-    },
-  },
-})
-
-export const printSectionHeadingRule = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "block",
-      width: "418px",
-      borderTop: `1px solid ${colors.text}`,
-      marginTop: vspace.quarter,
-    },
-  },
-})
-
-export const printBodyContent = style({
-  display: "none",
-  "@media": {
-    [mediaQueries.print]: {
-      display: "block",
-      width: "100%",
-      marginTop: 0,
-      breakBefore: "avoid",
-    },
-  },
-})
-
-export const printHideParagraphTranslation = style({
-  "@media": {
-    [mediaQueries.print]: { display: "block" },
-  },
-})
-
-globalStyle(`.${printHideParagraphTranslation} .${inlineBlock}`, {
-  "@media": {
-    [mediaQueries.print]: { display: "none" },
-  },
-})
-
-export const printWordSpacing = globalStyle(
-  `.${printBodyContent} .${wordGroup}, .${printBodyContent} .${wordGroupInline}`,
-  {
-    "@media": {
-      [mediaQueries.print]: {
-        display: "block",
-        marginRight: 0,
-        marginBottom: vspace.quarter,
-        paddingTop: 0,
-        paddingBottom: 0,
-        breakInside: "avoid",
-      },
-    },
-  }
-)
-
-globalStyle(
-  `.${printBodyContent} .${annotationSection.story} .${wordGroupInline}`,
-  {
-    "@media": {
-      [mediaQueries.print]: {
-        display: "block",
-        marginRight: 0,
-        marginBottom: vspace.quarter,
-      },
-    },
-  }
-)
-
-export const printDocumentBlockSpacing = globalStyle(
-  `.${printBodyContent} .${documentBlock.story}, .${printBodyContent} .${documentBlock.wordByWord}`,
-  {
-    "@media": {
-      [mediaQueries.print]: {
-        marginTop: vspace.quarter,
-        marginBottom: vspace.quarter,
-        paddingBottom: 0,
-        breakAfter: "avoid",
-        breakBefore: "avoid",
-      },
-    },
-  }
-)
-
-export const printAnnotationSectionSpacing = globalStyle(
-  `.${printBodyContent} .${annotationSection.story}, .${printBodyContent} .${annotationSection.wordParts}`,
-  {
-    "@media": {
-      [mediaQueries.print]: {
-        display: "grid",
-        gridTemplateColumns: "repeat(3, 212px)",
-        columnGap: vspace.one,
-        rowGap: vspace.one,
-        marginBottom: 0,
-      },
-    },
-  }
-)
-
-export const printFirstDocumentBlockBorder = globalStyle(
-  `.${printBodyContent} .${documentBlock.story}:first-child, .${printBodyContent} .${documentBlock.wordByWord}:first-child`,
-  {
-    "@media": {
-      [mediaQueries.print]: {
-        borderBottom: `1px solid ${colors.text}`,
-      },
-    },
-  }
-)
-
-export const printParagraphPageBreak = globalStyle(
-  `.${printBodyContent} .${wordGroup}:nth-child(3n+1):last-child, .${printBodyContent} .${wordGroupInline}:nth-child(3n+1):last-child`,
-  {
-    "@media": {
-      [mediaQueries.print]: {
-        breakAfter: "page",
-      },
-    },
-  }
-)
 
 export const printSyllabaryLine = globalStyle(
   `.${wordGroup} div[lang="chr"], .${wordGroupInline} div[lang="chr"]`,
